pyproject.toml
[tool.poetry]
name = "Schedule"
version = "0.1.0"
description = ""
authors = ["arsen2019 <arsen.grigoryan555@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
uvicorn = {extras = ["standart"], version = "^0.30.1"}
fastapi = "^0.111.0"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.30"}
pydantic-settings = "^2.3.1"
python-dotenv = "^1.0.1"


[tool.poetry.group.dev.dependencies]
black = "^24.4.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
